requirements.txt
fastapi
uvicorn
pandas
python-dateutil
requests
python-dotenv